CONNECT-321 Move common functionality into riak_backend
